test: 우산, 사물함 대여 모달 창 최적화 시도 중입니다
